Renaming for consistency with pages names.
