+++ components/trajectory/CertificateItem.tsx
