+++ components/trajectory/PathSkeleton.tsx
