+++ components/trajectory/StatusBadge.tsx
