+++ components/trajectory/TrajectoryDetailModal.tsx
