+++ components/trajectory/TrajectorySkeleton.tsx
