+++ components/trajectory/TrajectoryUtils.tsx
